requirements/base.txt
# Bleeding edge Django
django==1.7.1

# Configuration
django-configurations==0.8
django-secure==1.0
django-cache-url==0.8.0
dj-database-url==0.3.0

# Forms
django-braces==1.4.0
django-crispy-forms==1.4.0
django-floppyforms==1.2.0

# Models
django-model-utils==2.2

# images
Pillow==2.6.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.18.0

# For the persistance stores
psycopg2==2.5.4

# Unicode slugification
unicode-slugify==0.1.1
django-autoslug==1.7.2

# Useful things
django-avatar==2.0

# Time zones support
pytz==2014.7

# Your custom requirements go here
django-cms>=3.0

# These dependencies are brought in by django CMS, but if you want to
# lock-in their version, specify them

South==1.0
django-mptt==0.6
django-sekizai==0.7
django-classy-tags==0.5
djangocms-admin-style==0.2.2
html5lib==0.999
six==1.3.0

# Optional, recommended packages
# Pillow==2.0.0
django-filer==0.9.5
cmsplugin-filer==0.9.5
django-reversion==1.7
django-authtools
pysolr

#s-e git://github.com/Fantomas42/django-blog-zinnia.git@feature/django-1.7#egg=django-blog-zinnia